Verify target has all addresses found in the hex file instead of verifying the size.
As noted in PR #2512 simply verifying the size of the hex file is not
extensible and is not how hex files are designed to be used.

A hex file can contains many non-contiguous blocks of data.

This also supports targets with any partition layout.

Additionally, this provides a more flexible way of providing hex files
which can now contain data for non-contiguous blocks, e.g. firmware +
config or firmware + osd fonts.